Go back to whole reps
Removed after real use, and the reversal is recorded rather than quietly undone,
because the argument for halves is a good argument and somebody will make it
again.

What was true: a hand kept log really does carry halves, in about a quarter of
its sets, and a half really is the rep that got most of the way up.

What beat it: the reps stepper is the control touched most, one handed, under
fatigue, every set of every session. Halving the step doubles the taps for
everybody, permanently, to capture something that fits in the session note. The
general rule is that a cost paid every set outweighs information captured
occasionally, even when the information is real.

Existing halves are floored, not rounded to nearest, in both the Postgres
migration and the IndexedDB one. 10.5 is ten completed reps and a partial, so
rounding up would write down an eleventh rep nobody finished. Reads never
re-validate, so a stray half left on a device would have kept charting forever
while the schema said integers.

hold_seconds stays numeric. A hold that broke at 12.5 seconds is one observation
rather than half of anything, and it costs no taps: the stepper moves in fives
and a fraction only ever arrives by typing.

89 tests pass. The schema is now the guard: a fractional rep is rejected rather
than stored.

+++ js/schema.js
@@ -213,11 +213,12 @@ export const TABLES = {
       // count. Anything that computes volume or an estimated 1RM skips these rows rather than
       // inventing a number for them.
       //
-      // Numeric, not an integer, because a half rep is a real thing people write down: the rep
-      // that got most of the way up. Rounding it away throws out the difference between a
-      // session that improved and one that did not, and rounding it up invents a rep nobody
-      // completed. Epley and volume are both indifferent to it.
-      reps: { type: NUMERIC, nullable: true },
+      // Whole reps only. Halves were built, shipped, and used, and they did not survive contact
+      // with a real session: the argument for them was that a half is real information, and the
+      // argument that beat it is that doubling the taps on the one control used most under
+      // fatigue costs more than the information is worth. A partial rep goes in the session
+      // note, where it costs nothing to write and nothing to read.
+      reps: { type: INT, nullable: true },
       // Rounds completed in an AMRAP block. Null everywhere else.
       rounds: { type: INT, nullable: true },
       // How long a hold lasted, in seconds. An L sit and a hollow body have no reps and no
